chore: refactoring v1 and v2
