+++ pom.xml
@@ -322,17 +322,17 @@
                             <goal>generate</goal>
                         </goals>
                         <configuration>
-                            <inputSpec>${project.basedir}/api-spec/payment-requests-api.yaml</inputSpec>
+                            <inputSpec>${project.basedir}/api-spec/v1/payment-requests-api.yaml</inputSpec>
                             <generatorName>kotlin-spring</generatorName>
                             <library>spring-boot</library>
                             <modelNameSuffix>Dto</modelNameSuffix>
                             <generateApiTests>false</generateApiTests>
                             <generateModelTests>false</generateModelTests>
                             <configOptions>
-                                <basePackage>it.pagopa.ecommerce.generated.payment.requests.server</basePackage>
-                                <modelPackage>it.pagopa.ecommerce.generated.payment.requests.server.model</modelPackage>
-                                <apiPackage>it.pagopa.ecommerce.generated.payment.requests.server.api</apiPackage>
-                                <configPackage>it.pagopa.ecommerce.generated.payment.requests.server.config
+                                <basePackage>it.pagopa.ecommerce.generated.payment.requests.server.v1</basePackage>
+                                <modelPackage>it.pagopa.ecommerce.generated.payment.requests.server.v1.model</modelPackage>
+                                <apiPackage>it.pagopa.ecommerce.generated.payment.requests.server.v1.api</apiPackage>
+                                <configPackage>it.pagopa.ecommerce.generated.payment.requests.server.v1.config
                                 </configPackage>
                                 <interfaceOnly>true</interfaceOnly>
                                 <hideGenerationTimestamp>true</hideGenerationTimestamp>
@@ -353,7 +353,7 @@
                         </goals>
                         <configuration>
                             <!-- Path al nuovo OpenAPI v2 -->
-                            <inputSpec>${project.basedir}/api-spec/payment-requests-api-v2.yaml</inputSpec>
+                            <inputSpec>${project.basedir}/api-spec/v2/payment-requests-api.yaml</inputSpec>
                             <generatorName>kotlin-spring</generatorName>
                             <library>spring-boot</library>
                             <modelNameSuffix>Dto</modelNameSuffix>
@@ -374,7 +374,6 @@
                                 <useJakartaEe>true</useJakartaEe>
                                 <useSpringBoot3>true</useSpringBoot3>
                                 <useTags>true</useTags>
-                                <apiNameSuffix>V2</apiNameSuffix>
                             </configOptions>
                         </configuration>
                     </execution>
