+++ src/main/kotlin/it/pagopa/ecommerce/payment/requests/controllers/PaymentRequestsController.kt
@@ -1,7 +1,7 @@
 package it.pagopa.ecommerce.payment.requests.controllers
 
-import it.pagopa.ecommerce.generated.payment.requests.server.api.PaymentRequestsApi
-import it.pagopa.ecommerce.generated.payment.requests.server.model.PaymentRequestsGetResponseDto
+import it.pagopa.ecommerce.generated.payment.requests.server.v1.api.PaymentRequestsApi
+import it.pagopa.ecommerce.generated.payment.requests.server.v1.model.PaymentRequestsGetResponseDto
 import it.pagopa.ecommerce.payment.requests.services.PaymentRequestsService
 import it.pagopa.ecommerce.payment.requests.warmup.annotations.WarmupFunction
 import it.pagopa.ecommerce.payment.requests.warmup.exceptions.WarmUpException
