+++ src/main/kotlin/it/pagopa/ecommerce/payment/requests/repositories/redistemplate/v2/CartsRedisTemplateWrapper.kt
@@ -1,0 +1,15 @@
+package it.pagopa.ecommerce.payment.requests.repositories.redistemplate.v2
+
+import it.pagopa.ecommerce.payment.requests.repositories.redistemplate.ReactiveRedisTemplateWrapper
+import it.pagopa.ecommerce.payment.requests.repositories.v2.CartInfo
+import java.time.Duration
+import org.springframework.data.redis.core.ReactiveRedisTemplate
+
+class CartsRedisTemplateWrapper(
+  reactiveRedisTemplate: ReactiveRedisTemplate<String, CartInfo>,
+  ttl: Duration
+) :
+  ReactiveRedisTemplateWrapper<CartInfo>(
+    reactiveRedisTemplate = reactiveRedisTemplate, "carts", ttl) {
+  override fun getKeyFromEntity(value: CartInfo): String = value.id.toString()
+}
