+++ src/test/kotlin/it/pagopa/ecommerce/payment/requests/configurations/redis/RedisConfigurationTest.kt
@@ -9,10 +9,10 @@ import it.pagopa.ecommerce.payment.requests.configurations.redis.converters.Jack
 import it.pagopa.ecommerce.payment.requests.configurations.redis.converters.JacksonRptSerializer
 import it.pagopa.ecommerce.payment.requests.domain.IdempotencyKey
 import it.pagopa.ecommerce.payment.requests.domain.RptId
-import it.pagopa.ecommerce.payment.requests.repositories.CartInfo
 import it.pagopa.ecommerce.payment.requests.repositories.PaymentInfo
 import it.pagopa.ecommerce.payment.requests.repositories.PaymentRequestInfo
-import it.pagopa.ecommerce.payment.requests.repositories.ReturnUrls
+import it.pagopa.ecommerce.payment.requests.repositories.v1.CartInfo
+import it.pagopa.ecommerce.payment.requests.repositories.v1.ReturnUrls
 import java.nio.ByteBuffer
 import java.util.*
 import org.junit.jupiter.api.Assertions.assertEquals
